+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.4.1] - 2019-10-17
 ### Added
 - Added `containerDepth` to avifDecoder for surfacing 10bpc/12bpc flags from av1C boxes, if present
 - Added `avifCodecVersions()` for getting version strings of internal AV1 codecs
@@ -191,7 +193,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Constants `AVIF_VERSION`, `AVIF_VERSION_MAJOR`, `AVIF_VERSION_MINOR`, `AVIF_VERSION_PATCH`
 - `avifVersion()` function
 
-[Unreleased]: https://github.com/AOMediaCodec/libavif/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/AOMediaCodec/libavif/compare/v0.4.1...HEAD
+[0.4.1]: https://github.com/AOMediaCodec/libavif/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/AOMediaCodec/libavif/compare/v0.3.11...v0.4.0
 [0.3.11]: https://github.com/AOMediaCodec/libavif/compare/v0.3.10...v0.3.11
 [0.3.10]: https://github.com/AOMediaCodec/libavif/compare/v0.3.9...v0.3.10